second commit for homework2
